# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social Media Platform",
        "title": "Social Media API"
    },
    "servers": [
        {
            "url": "http://localhost:3200"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": ["Posts"],
                "summary": "Get All Posts",
                "description": "Retrieve all posts from the database",
                "security": [{"JWT": {}}],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": ["Posts"],
                "summary": "Create a Post",
                "description": "Create a new post",
                "security": [{"JWT": {}}],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "integer"
                                    },
                                    "caption": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    }
}